Refactor Hotspot
